# src/main/resources/integration/epo/openapi.yaml
openapi: 3.0.1
info:
  title: Woxy NPO Quality API
  description: API of the Woxy
  termsOfService: https://www.epo.org
  contact:
    name: European Patent Office
    url: https://www.epo.org
    email: info@epo.org
  license:
    name: Commercial
  version: 1.0.0
servers:
- url: https://iodq.apps-staging.epo.org/dataquality/data-quality-service-intg
  description: Generated server url
paths:
...
  /api/office-logo/{countryCode}:
    get:
      tags:
      - NpoOfficeLogo
      summary: Download logo of office with provided country code
      operationId: download_3
      parameters:
      - name: countryCode
        in: path
        description: ISO 3166 2-letter country code
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Unauthorized
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - NpoOfficeLogo
      summary: Upload file to be used for validation of the structure of XML PublicationFiles
        per country and data type
      operationId: upload_2
      parameters:
      - name: countryCode
        in: path
        description: ISO 3166 2-letter country code
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  description: File to be uploaded
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
              example: grOfficeLogo.png
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Unauthorized
        "405":
          description: Method Not Allowed
        "415":
          description: Unsupported Media Type
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/user-guide/download:
    get:
      tags:
      - UserGuide
      summary: Download latest user-guide file
      operationId: download
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Unauthorized
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/structure-validation/file/{countryCode}/{dataType}/download:
    get:
      tags:
      - NpoStructureValidationFile
      summary: Download file being used for validation of the structure of XML files
      operationId: download_1
      parameters:
      - name: countryCode
        in: path
        description: ISO 3166 2-letter country code
        required: true
        schema:
          type: string
      - name: dataType
        in: path
        description: PublicationFile's DataType for which the file is being used for
          validation
        required: true
        schema:
          type: string
          enum:
          - ABSTRACT
          - AUTHORITY_FILE
          - BIBLIOGRAPHIC
          - CITATION
          - COMBINED_DATA_PACKAGE
          - CPC
          - FULL_TEXT
          - IMAGE
          - LEGAL_STATUS
          - UNKNOWN
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Unauthorized
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/publicationFile/download/{publicationFileId}:
    get:
      tags:
      - NpoPublicationFile
      summary: Download Publication File based on its ID
      operationId: download_2
      parameters:
      - name: publicationFileId
        in: path
        description: ID of the PublicationFile
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: binary
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Unauthorized
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      required:
      - errors
      type: object
      properties:
        errors:
          type: array
          description: Includes all errors caused by a request
          items:
            $ref: '#/components/schemas/ApiErrorDetail'
    ApiErrorDetail:
      required:
      - code
      - title
      type: object
      properties:
        code:
          type: string
          description: Unique code of the error
        title:
          type: string
          description: Title of error
        message:
          type: string
          description: Message of the error
        detail:
          type: string
          description: Details of the error
        source:
          $ref: '#/components/schemas/ErrorSource'
      description: Includes all errors caused by a request
    ErrorSource:
      required:
      - pointer
      type: object
      properties:
        pointer:
          type: string
          description: Points to the location of the request caused this issue.
      description: Inidate path of the input property caused the error
...